Refactor Post and NewPost layouts into Post
