+++ client/src/pages/Posts.js
@@ -45,7 +45,7 @@ export default function Posts({ authUser }) {
       {
         posts.map(post =>
           post._id === "new" ?
-            <NewPost key={post._id} setPosts={setPosts} authUser={authUser} />
+            <Post isNew={true} key={post._id} setPosts={setPosts} authUser={authUser} />
           :
             <Post key={post._id} post={post} setPosts={setPosts} authUser={authUser} />
         )
